Change title management approach
