+++ src/components/MediatedRequestsActivities/MediatedRequestsActivities.js
@@ -6,16 +6,13 @@ import {
 } from 'react-intl';
 import {
   useHistory,
-  useLocation,
 } from 'react-router-dom';
 import { cloneDeep } from 'lodash';
-import { parse } from 'query-string';
 
 import {
   IfPermission,
   AppIcon,
   useCallout,
-  TitleManager,
 } from '@folio/stripes/core';
 import {
   Button,
@@ -228,7 +225,6 @@ const MediatedRequestsActivities = ({
 }) => {
   const [filterPaneIsVisible, setFilterPaneIsVisible] = useState(true);
   const history = useHistory();
-  const location = useLocation();
   const callout = useCallout();
   const [isLoadingReport, setIsLoadingReport] = useState(false);
   const { formatMessage } = useIntl();
@@ -254,118 +250,101 @@ const MediatedRequestsActivities = ({
     );
   };
 
-  const getPageTitle = () => {
-    const query = parse(location.search)?.query;
-
-    if (query) {
-      return formatMessage(
-        { id: 'ui-requests-mediated.meta.searchTitle' },
-        { query },
-      );
-    }
-
-    return formatMessage({ id: 'ui-requests-mediated.meta.title' });
-  };
-
   const mediatedRequests = resources[MEDIATED_REQUESTS_RECORDS_NAME]?.records ?? [];
   const query = queryGetter ? queryGetter() || {} : {};
-  const pageTitle = getPageTitle();
 
   return (
-    <>
-      <TitleManager page={pageTitle} />
-      <SearchAndSortQuery
-        data-testid="mediatedRequestsActivitiesSearchAndSortQuery"
-        initialSearchState={{ query: '' }}
-        syncToLocationSearch={false}
-        querySetter={querySetter}
-        queryGetter={queryGetter}
-      >
-        {({
-          activeFilters,
-          getSearchHandlers,
-          getFilterHandlers,
-          searchValue,
-          onSort,
-          onSubmitSearch,
-          resetAll,
-        }) => (
-          <Paneset data-testid="mediatedRequestsActivitiesPaneSet">
-            {filterPaneIsVisible &&
-            <Pane
-              data-testid="mediatedRequestsActivitiesPane"
-              defaultWidth={FILTER_PANE_WIDTH}
-              paneTitle={<FormattedMessage id="ui-requests-mediated.app.filterPane.selectActivity" />}
-              lastMenu={
-                <PaneMenu>
-                  <CollapseFilterPaneButton
-                    data-testid="mediatedRequestsActivitiesCollapseFilterPaneButton"
-                    onClick={toggleFilterPane}
-                  />
-                </PaneMenu>
-              }
-            >
-              <NavigationMenu
-                value={getMediatedRequestsActivitiesUrl()}
-                separator
-              />
-              <MediatedRequestsFilters
-                activeFilters={activeFilters?.state}
-                getSearchHandlers={getSearchHandlers}
-                searchValue={searchValue}
-                onSubmitSearch={onSubmitSearch}
-                resetAll={resetAll}
-                onChangeHandlers={getFilterHandlers()}
-                resultOffset={resultOffset}
-                settings={settings}
-              />
-            </Pane>
+    <SearchAndSortQuery
+      data-testid="mediatedRequestsActivitiesSearchAndSortQuery"
+      initialSearchState={{ query: '' }}
+      syncToLocationSearch={false}
+      querySetter={querySetter}
+      queryGetter={queryGetter}
+    >
+      {({
+        activeFilters,
+        getSearchHandlers,
+        getFilterHandlers,
+        searchValue,
+        onSort,
+        onSubmitSearch,
+        resetAll,
+      }) => (
+        <Paneset data-testid="mediatedRequestsActivitiesPaneSet">
+          {filterPaneIsVisible &&
+          <Pane
+            data-testid="mediatedRequestsActivitiesPane"
+            defaultWidth={FILTER_PANE_WIDTH}
+            paneTitle={<FormattedMessage id="ui-requests-mediated.app.filterPane.selectActivity" />}
+            lastMenu={
+              <PaneMenu>
+                <CollapseFilterPaneButton
+                  data-testid="mediatedRequestsActivitiesCollapseFilterPaneButton"
+                  onClick={toggleFilterPane}
+                />
+              </PaneMenu>
             }
-            <ColumnManager
-              id="mediatedRequestsActivitiesColumnManager"
-              columnMapping={MEDIATED_REQUESTS_RECORD_TRANSLATIONS}
-              excludeKeys={[MEDIATED_REQUESTS_RECORD_FIELD_NAME.TITLE]}
-            >
-              {({ renderColumnsMenu, visibleColumns }) => (
-                <Pane
-                  defaultWidth="fill"
-                  padContent={false}
-                  noOverflow
-                  appIcon={<AppIcon app={APP_ICON_NAME} />}
-                  paneTitle={<FormattedMessage id="ui-requests-mediated.mediatedRequestsActivities.paneTitle" />}
-                  paneSub={getResultPaneSub(source)}
-                  firstMenu={renderResultsFirstMenu(activeFilters)}
-                  actionMenu={
-                    getActionMenu({
-                      renderColumnsMenu,
-                      mediatedRequests,
-                      searchValue,
-                      activeFilters,
-                      reportRecords,
-                      callout,
-                      history,
-                      isLoadingReport,
-                      setIsLoadingReport,
-                      formatMessage,
-                    })
-                  }
-                >
-                  <MediatedRequestsList
-                    visibleColumns={visibleColumns}
-                    contentData={mediatedRequests}
-                    source={source}
-                    query={query}
-                    onSort={onSort}
-                    onNeedMoreData={onNeedMoreData}
-                  />
-                </Pane>
-              )}
-            </ColumnManager>
-            { children }
-          </Paneset>
-        )}
-      </SearchAndSortQuery>
-    </>
+          >
+            <NavigationMenu
+              value={getMediatedRequestsActivitiesUrl()}
+              separator
+            />
+            <MediatedRequestsFilters
+              activeFilters={activeFilters?.state}
+              getSearchHandlers={getSearchHandlers}
+              searchValue={searchValue}
+              onSubmitSearch={onSubmitSearch}
+              resetAll={resetAll}
+              onChangeHandlers={getFilterHandlers()}
+              resultOffset={resultOffset}
+              settings={settings}
+            />
+          </Pane>
+          }
+          <ColumnManager
+            id="mediatedRequestsActivitiesColumnManager"
+            columnMapping={MEDIATED_REQUESTS_RECORD_TRANSLATIONS}
+            excludeKeys={[MEDIATED_REQUESTS_RECORD_FIELD_NAME.TITLE]}
+          >
+            {({ renderColumnsMenu, visibleColumns }) => (
+              <Pane
+                defaultWidth="fill"
+                padContent={false}
+                noOverflow
+                appIcon={<AppIcon app={APP_ICON_NAME} />}
+                paneTitle={<FormattedMessage id="ui-requests-mediated.mediatedRequestsActivities.paneTitle" />}
+                paneSub={getResultPaneSub(source)}
+                firstMenu={renderResultsFirstMenu(activeFilters)}
+                actionMenu={
+                  getActionMenu({
+                    renderColumnsMenu,
+                    mediatedRequests,
+                    searchValue,
+                    activeFilters,
+                    reportRecords,
+                    callout,
+                    history,
+                    isLoadingReport,
+                    setIsLoadingReport,
+                    formatMessage,
+                  })
+                }
+              >
+                <MediatedRequestsList
+                  visibleColumns={visibleColumns}
+                  contentData={mediatedRequests}
+                  source={source}
+                  query={query}
+                  onSort={onSort}
+                  onNeedMoreData={onNeedMoreData}
+                />
+              </Pane>
+            )}
+          </ColumnManager>
+          { children }
+        </Paneset>
+      )}
+    </SearchAndSortQuery>
   );
 };
