+++ src/components/MediatedRequestsActivities/MediatedRequestsActivities.test.js
@@ -1,7 +1,6 @@
 import {
   MemoryRouter,
   useHistory,
-  useLocation,
 } from 'react-router-dom';
 
 import {
@@ -14,7 +13,6 @@ import {
   exportToCsv,
   filters2cql,
 } from '@folio/stripes/components';
-import { TitleManager } from '@folio/stripes/core';
 
 import MediatedRequestsActivities, {
   getActionMenu,
@@ -44,8 +42,6 @@ const labelIds = {
   reportPendingButton: 'ui-requests-mediated.mediatedRequestList.actionMenu.reportPending',
   searchCriteria: 'stripes-smart-components.searchCriteria',
   searchResultsCount: 'stripes-smart-components.searchResultsCountHeader',
-  pageTitle: 'ui-requests-mediated.meta.title',
-  searchTitle: 'ui-requests-mediated.meta.searchTitle',
 };
 
 jest.mock('@folio/stripes/smart-components', () => ({
@@ -65,9 +61,6 @@ jest.mock('react-router-dom', () => ({
   ...jest.requireActual('react-router-dom'),
   useLocation: jest.fn(),
   useHistory: jest.fn(),
-}));
-jest.mock('query-string', () => ({
-  parse: jest.fn(search => search),
 }));
 jest.mock('../../utils', () => ({
   ...jest.requireActual('../../utils'),
@@ -104,14 +97,7 @@ describe('MediatedRequestsActivities', () => {
   };
 
   describe('When search term is entered', () => {
-    const query = 'searchQuery';
-
     beforeEach(() => {
-      useLocation.mockReturnValueOnce({
-        search: {
-          query,
-        },
-      });
       useHistory.mockReturnValueOnce({ push: jest.fn() });
 
       render(
@@ -121,12 +107,6 @@ describe('MediatedRequestsActivities', () => {
           />
         </MemoryRouter>
       );
-    });
-
-    it('should trigger TitleManager with correct props', () => {
-      expect(TitleManager).toHaveBeenCalledWith({
-        page: labelIds.searchTitle,
-      }, {});
     });
 
     it('should render search and sort query', () => {
@@ -155,26 +135,6 @@ describe('MediatedRequestsActivities', () => {
       expect(MediatedRequestsFilters).toHaveBeenCalledWith(expect.objectContaining({
         settings,
       }), {});
-    });
-  });
-
-  describe('When search term is not entered', () => {
-    beforeEach(() => {
-      useLocation.mockReturnValueOnce({
-        search: '',
-      });
-
-      render(
-        <MediatedRequestsActivities
-          {...basicProps}
-        />
-      );
-    });
-
-    it('should trigger TitleManager with correct props', () => {
-      expect(TitleManager).toHaveBeenCalledWith({
-        page: labelIds.pageTitle,
-      }, {});
     });
   });
 });
