+++ src/index.js
@@ -1,4 +1,7 @@
 import PropTypes from 'prop-types';
+import { useLocation } from 'react-router-dom';
+import { useIntl } from 'react-intl';
+import { parse } from 'query-string';
 import 'core-js/stable';
 import 'regenerator-runtime/runtime';
 
@@ -7,6 +10,7 @@ import {
   Switch,
   Redirect,
   useStripes,
+  TitleManager,
 } from '@folio/stripes/core';
 
 import ConfirmItemArrival from './components/ConfirmItemArrival';
@@ -30,66 +34,84 @@ import {
 
 const RequestsMediated = (props) => {
   const stripes = useStripes();
+  const location = useLocation();
+  const { formatMessage } = useIntl();
   const { settings } = useGeneralTlrSettings(!props.showSettings);
   const { patronGroups } = usePatronGroups(!props.showSettings);
+  const getPageTitle = () => {
+    const query = parse(location.search)?.query;
+
+    if (query) {
+      return formatMessage(
+        { id: 'ui-requests-mediated.meta.searchTitle' },
+        { query },
+      );
+    }
+
+    return formatMessage({ id: 'ui-requests-mediated.meta.title' });
+  };
 
   const {
     match: {
       path,
     },
   } = props;
+  const pageTitle = getPageTitle();
   const isRequestCreationAvailable = stripes.hasPerm('ui-requests-mediated.all') || stripes.hasPerm('ui-requests-mediated.view-create-edit');
 
   return (
-    <Switch>
-      {
-        isRequestCreationAvailable &&
+    <>
+      <TitleManager page={pageTitle} />
+      <Switch>
+        {
+          isRequestCreationAvailable &&
           <Route path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/create`}>
             <RequestFormContainer
               settings={settings}
               patronGroups={patronGroups?.usergroups}
             />
           </Route>
-      }
-      <Route path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/preview/:id`}>
-        <MediatedRequestsActivitiesContainer
-          {...props}
-          settings={settings}
-        >
-          <MediatedRequestsDetail
+        }
+        <Route path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/preview/:id`}>
+          <MediatedRequestsActivitiesContainer
             {...props}
-            patronGroups={patronGroups?.usergroups}
-          />
-        </MediatedRequestsActivitiesContainer>
-      </Route>
-      <Route
-        path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/new`}
-        component={NoteCreator}
-      />
-      <Route
-        path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/:noteId/edit`}
-        component={NoteEditor}
-      />
-      <Route
-        path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/:noteId`}
-        component={NoteViewer}
-      />
-      <Route path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}`}>
-        <MediatedRequestsActivitiesContainer
-          {...props}
-          settings={settings}
+            settings={settings}
+          >
+            <MediatedRequestsDetail
+              {...props}
+              patronGroups={patronGroups?.usergroups}
+            />
+          </MediatedRequestsActivitiesContainer>
+        </Route>
+        <Route
+          path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/new`}
+          component={NoteCreator}
         />
-      </Route>
-      <Route
-        path={`${path}/${CONFIRM_ITEM_ARRIVAL}`}
-        component={ConfirmItemArrival}
-      />
-      <Route
-        path={`${path}/${SEND_ITEM_IN_TRANSIT}`}
-        component={SendItemInTransit}
-      />
-      <Redirect to={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}`} />
-    </Switch>
+        <Route
+          path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/:noteId/edit`}
+          component={NoteEditor}
+        />
+        <Route
+          path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}/notes/:noteId`}
+          component={NoteViewer}
+        />
+        <Route path={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}`}>
+          <MediatedRequestsActivitiesContainer
+            {...props}
+            settings={settings}
+          />
+        </Route>
+        <Route
+          path={`${path}/${CONFIRM_ITEM_ARRIVAL}`}
+          component={ConfirmItemArrival}
+        />
+        <Route
+          path={`${path}/${SEND_ITEM_IN_TRANSIT}`}
+          component={SendItemInTransit}
+        />
+        <Redirect to={`${path}/${MEDIATED_REQUESTS_ACTIVITIES}`} />
+      </Switch>
+    </>
   );
 };
